Make it attachable to an Exposure
This is only meant as a temporary solution. Eventually, we would like to
stitch the cell-based coadds on the fly and run detection, deblending,
measurement etc. But in the short run, it is useful to save the stitched
coadd as an Exposure and used in other tasks. The fix to make it an
lsst.afw.image.ApCorrMap is definitely not clean, but it works for now
and is meant only as a crutch. When this crutch is no longer required,
it should be possible to simply revert this commit.