Allow for padding when constructing
MultipleCellCoadd instances. With DM-46961, the cells at the boundary
can have rectangular bounding boxes to cover the entire patch. This will
neither correspond to the inner bbox nor the outer bbox. To accommodate
for this, the check at the time of construction has to be loosened to
check that the bounding box contains inner bbox and is contained within
the outer bbox.

Co-authored-by: Jim Bosch <jbosch@astro.princeton.edu>